+++ app/src/renderer/components/Organisation/OrganisationOverview/OrganisationList.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script>
-    import Organisation from './Organization';
+    import Organisation from './Organisation';
 
     let organizations = [];
